Crate hoover3_tracing

Tracing and logging initialization - used by both frontend and backend.

Functions

_init_logging_2 🔒

Experiment into manually configuring logging. Currenti issues: - tracing_wasm (which is used by dioxus_logger) is old/unmaintained and does not allow configuration of message format (only turning colors on/off). - tracing_web, a different crate, is not compatible with dioxus-cli + fullstack (logs don’t show up in terminal). - other crates I tried cause some kind of panic in the browser. - https://github.com/old-storyai/tracing-wasm/issues/30 - https://github.com/block-mesh/block-mesh-monorepo/blob/master/libs/logger-leptos/src/leptos_tracing.rs

init_logging 🔒

init_tracing

Initialize tracing and log crates, using dioxus_logger TODO: distributed tracing (SigNoz).